requirements.txt
strands-agents
strands-agents[ollama]
fastmcp
